packages/api/requirements.txt
alembic==0.9.9
astroid==2.2.5
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
click==6.7
Faker==0.8.15
Flask==0.12.2
Flask-Cors==3.0.4
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.8.1
idna==2.8
isort==4.3.18
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.4.0
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.5.0
PyJWT==1.6.1
pylint==2.3.1
python-dateutil==2.7.3
python-editor==1.0.3
requests==2.21.0
six==1.11.0
SQLAlchemy==1.2.7
stripe==2.27.0
text-unidecode==1.2
typed-ast==1.3.5
urllib3==1.24.3
Werkzeug==0.14.1
wrapt==1.11.1
